// LinkERP/ClientApp/src/app/models/pur/tendertypes.ts
import { BaseEntity } from "../base-entity";

export class Tendertypes extends BaseEntity {
  Description: any;
  FastKey: any;
  Caption1: any;
  Caption2: any;
  Caption3: any;
  Caption4: any;
  Caption5: any;
  Caption6: any;
  Caption7: any;
  Caption8: any;
  Caption1IsRequired: any;
  Caption2IsRequired: any;
  Caption3IsRequired: any;
  Caption4IsRequired: any;
  Caption5IsRequired: any;
  Caption6IsRequired: any;
  Caption7IsRequired: any;
  Caption8IsRequired: any;
  Caption1FieldType: any;
  Caption2FieldType: any;
  Caption3FieldType: any;
  Caption4FieldType: any;
  Caption5FieldType: any;
  Caption6FieldType: any;
  Caption7FieldType: any;
  Caption8FieldType: any;
  Caption2List: any;
  Caption3List: any;
  Caption4List: any;
  Caption5List: any;
  Caption6List: any;
  Caption1List: any;
  Caption7List: any;
  Caption8List: any;
  VerifyFromListCaption1: any;
  VerifyFromListCaption2: any;
  VerifyFromListCaption3: any;
  VerifyFromListCaption4: any;
  VerifyFromListCaption5: any;
  VerifyFromListCaption6: any;
  VerifyFromListCaption7: any;
  VerifyFromListCaption8: any;
  SurchargePercentValue: any;
  SurchargeAmount: any;
  DiscountPercentValue: any;
  DiscountAmount: any;
  MultiCurrency: any;
  EFT: any;
  IncludeInBanking: any;
  BankAccountCode: any;
  ShiftVarianceAccountCode: any;
  CompanyID: any;
  TenderTypesGroup: any;




}